msp430-as : step 5 : Implement support for control structures
I noticed a spelling mistake in the template that went unnoticed in quite
a few commits of the previous ports. Since I'm not too sure how I can go
back in time to fix all of that in git, I'll leave them be, since from now
on at least, newer ports will not carry over that mistake.