Modify the test at step 3 to catch bugs in emit/key early on
For ease of testing, key/emit need to be implemented before dup/drop and
then they need to be modified to use dup/drop respectively after dup/drop
have been implemented. If you forget to make the mods for the second part,
the bug is caught much later, usually at step 27, by which time is it much
harder to figure out what is going on.

The purpose of this change is to catch this bug as early as possible when
the code changes for key and emit are still fresh in your mind.