msp430-as : step 33 : Verify signum(~) works
While making the changes related to enter/exit, I mistakenly thought that
"exit" was reserved by the msp430-gcc compiler for its own use since I saw
it in the objdump output. So I renamed "exit" to "return" back then. But at
this step, "exit" is used by the tests in rom.fth/defs.fth so I decided to
revert that change and it seemingly works without any name clashes.

Using "return" will not fly either since it clashes with the C reserved
word when used in the C port so I'll stick with using "exit" until there is
a clear sign that the current naming scheme causes problems.