msp430-as : step 36 : Implement 0=
Implementing this overflowed the ROM by 2 bytes so I reorganized the code
a bit to save 2 bytes. The savings comes from using the @reg+ address mode
but this needed the return stack to grow to low memory.

As of this commit, the ROM is full so rather than try space optimizing for
slim pickings, I'm going to use a larger device as the compile target. I'll
defer that change to the next commit.