Fix a (previously existing) regression testing failure
Running a full regression test was failing for x86-32 while running
	./runallsteps x86-32
This was because zig was erroring out when constants were too large
to fit rather than just warn about it (which is the default on all the
other compilers). So I decided to just go ahead and fix it by adding the
special cases for the constant values based on CELL width.

I'm not sure if this was due to upgrading zig to the latest 0.11.0 version