msp430-stc : step 57 : Enable conditionals at the repl
In the offset==1 code path within step 57 in rest.inp, I've decided to
leave the one remaining 'lit' unchanged while everything else has been
replaced with pad0 since making that change will break the x86 and x86-as
builds. I plan to fix it RSN but not as part of this commit since it is
already a little bit on the largish side.