pdk14-thr1 : workaround the regression due to increased ROM size
Due to the additional stack sanity checks added in steps 5, 17 and 18
the size of the stuff that needs to be in the ROM had increased enough
to cause trouble for the pdk-thr1 implementation which was barely able
to hold itself together to get past step 31. Rather than try to fix it
I'm disabling the tests for pdk-thr1 since there is no reason to touch
it - any regressions will hopefully be caught by the other ports.